vllm_deployment_final/requirements.txt
# vLLM inference server
vllm
fastapi
uvicorn
pydantic

# LangChain RAG pipeline
langchain
langchain-community
langchain-huggingface

# LlamaIndex RAG pipeline
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-readers-file

# Shared: vector store and embeddings
chromadb
sentence-transformers

# Environment
python-dotenv

# Benchmark dependencies
psutil
torch
numpy
matplotlib
aiohttp
datasets
